% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Class.R
\docType{class}
\name{GBFeature-class}
\alias{GBFeature-class}
\title{GenBank Feature Class}
\description{
S4 class representing a GenBank feature with type, location, and qualifiers
}
\section{Slots}{

\describe{
\item{\code{type}}{Character specifying feature type (e.g., "CDS", "gene", "mRNA")}

\item{\code{location}}{Character describing genomic location (e.g., "1..500")}

\item{\code{qualifiers}}{List of key-value pairs with feature attributes}
}}

\author{
Sergej Ruff
}
\keyword{internal}
