% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Class.R
\docType{class}
\name{GBReference-class}
\alias{GBReference-class}
\title{GenBank Reference Class}
\description{
S4 class representing a GenBank reference entry
}
\section{Slots}{

\describe{
\item{\code{number}}{Character reference index}

\item{\code{authors}}{Character vector of authors}

\item{\code{title}}{Character reference title}

\item{\code{journal}}{Character journal information}

\item{\code{pubmed}}{Character PubMed ID}
}}

\author{
Sergej Ruff
}
\keyword{internal}
