% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Class.R
\docType{class}
\name{GBSequence-class}
\alias{GBSequence-class}
\title{GenBank Sequence Class}
\description{
S4 class representing a complete GenBank record
}
\section{Slots}{

\describe{
\item{\code{LOCUS}}{Character locus information}

\item{\code{DEFINITION}}{Character sequence definition}

\item{\code{ACCESSION}}{Character accession number}

\item{\code{VERSION}}{Character version information}

\item{\code{DBLINK}}{Character database links}

\item{\code{KEYWORDS}}{Character keywords}

\item{\code{SOURCE}}{Character source organism}

\item{\code{ORGANISM}}{Character organism taxonomy}

\item{\code{REFERENCE}}{List of GBReference objects}

\item{\code{COMMENT}}{Character comments}

\item{\code{FEATURES}}{List of GBFeature objects}

\item{\code{ORIGIN}}{Character DNA sequence}
}}

\author{
Sergej Ruff
}
\keyword{internal}
