% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_Class.R
\docType{class}
\name{GenBankyObj-class}
\alias{GenBankyObj-class}
\title{GenBank Collection Class}
\description{
S4 class containing multiple GenBank records
}
\details{
Contains attribute "changed_accessions" showing any modified accession names
when duplicates were present. Access with \code{attr(object, "changed_accessions")}
}
\section{Slots}{

\describe{
\item{\code{sequences}}{List of GBSequence objects, named by accession numbers}
}}

\author{
Sergej Ruff (modified from geneviewer package)
}
\keyword{internal}
